Renamed parser.py to cc_rules_current to better reflect the file's actual function
